+++ generator-lateralus/component/templates/view.js
@@ -13,14 +13,17 @@ define([
 ) {
   'use strict';
 
-  var {{componentClassName}}ComponentView = Lateralus.Component.View.extend({
+  var Base = Lateralus.Component.View;
+  var baseProto = Base.prototype;
+
+  var {{componentClassName}}ComponentView = Base.extend({
     template: template
 
     /**
      * @param {Object} [options] See http://backbonejs.org/#View-constructor
      */
     ,initialize: function () {
-      this._super('initialize', arguments, {{componentClassName}}ComponentView);
+      baseProto.initialize.apply(this, arguments);
     }
   });